move `AuthenticatedEncryptionError` to the `errors` module
